+++ example/lib/main.dart
@@ -97,7 +97,7 @@ class ExampleViewer extends StatelessWidget {
                 height: constraints.maxHeight,
               ),
               htmlViewOptions: const HtmlViewOptions(
-                contentClass: 'sample-content',
+                contentClassName: 'sample-content',
               ),
               topBarOptions: TopBarOptions(
                 title: 'Some title',